requirements.txt
onnxruntime>=1.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
pillow>=10.0.0
flask==2.2.3
flask-cors==3.0.10
python-dotenv>=1.0.0
uvicorn>=0.15.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
cryptography>=41.0.0
transformers>=4.35.0
torch>=2.1.0
Werkzeug==2.2.3
tf2onnx>=1.13.0
tqdm>=4.66.0 